feat(orbital): major refactor of script
This adjusts a lot of the guardrails around how we cleanup if the
commands fail in some way. After a lot of thinking about how to handle
the PIPE requires some type of content even if it is empty to exit
cleanly. I decided to use the `echo "" > PIPE` to solve this. Its not
perfect but it does work successfully for Terraform because it wants a
`yes` as the only step to move forward.

Also the filter is moved to lookup `setup.tf` and we now pass `-print0`
so that we don't break the pipes.